fix: sycamore float64, live view timeout, debug zip timestamp, button style
solver_proc: sycamore centroid array dtype float32 → float64.  The Rust
PyO3 binding for solve_from_centroids requires float64; float32 caused
immediate pattern-match failure (status 2, 0 matches) regardless of
star count.

app.py: frame_jpg status call timeout 15s → 2s.  The default 15s
maint-socket timeout blocked each /live/frame.jpg request for up to 15s
while the daemon was starting or unavailable, making the live view
appear hung on first load.

dashboard.html: build debug ZIP filename from a JS timestamp instead
of parsing the Content-Disposition header.  Flask's header encoding
varied across versions and the regex fallback produced the bare
'diofinder_debug.zip' name with no timestamp.

dashboard.html: "Reset to center" and "Collect debug bundle" changed
from .btn to .btn.btn-primary so the #cc0000 border makes them visually
distinct from surrounding text.

+++ efinder/solver_proc.py
@@ -412,8 +412,9 @@ def solver_main(slots, latest_solution, shared_cfg,
                     )
                     # sycamore returns (x=col, y=row, brightness, peak).
                     # tetra3 solve_from_centroids expects (row, col) = (y, x).
+                    # Must be float64: Rust PyO3 binding rejects float32.
                     centroids = (
-                        np.array([[s[1], s[0]] for s in _raw], dtype=np.float32)
+                        np.array([[s[1], s[0]] for s in _raw], dtype=np.float64)
                         if _raw else None
                     )
                 else:
